add gitignore to git

+++ mkdircdgit.fish
@@ -2,8 +2,9 @@ function mkdircdgit --description 'make a directory, if it does not yet exist, c
   mkdircd "$pth" $argv && git init .
   if [ -n "$lang" ]
     set gt (fish_home)/gitignores/$lang.gitignore
-    if [ -e "$gt" ]
+    if [ -e "$gt" -a ! -e .gitignore ]
       cp "$gt" .gitignore
+      git add .gitignore
     end
   end
 end
